homework task5
