+++ js/data.js
@@ -1,0 +1,16 @@
+const TITLE = ['Отели', 'Апартаменты/квартиры', 'Курортные отели', 'Виллы', 'Шале', 'Коттеджи', 'Глэмпинг']; // заголовок предложения
+const PRICE = [20000, 40000, 50000, 70000, 150000, 32000, 90000]; // Рандом цен
+const TYPE = ['palace', 'flat', 'house', 'bungalow']; // одно из четырёх фиксированных значений
+const TIME = ['12:00', '13:00', '14:00']; //  одно из трёх фиксированных значений
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner']; // преимущества
+const DESCRIPTION = ['панорамным видом на Средиземное море оснащены',
+  'кондиционером', 'телевизором с плоским экраном и международными каналами',
+  'В каждом номере можно воспользоваться принадлежностями для чая/кофе.',
+  'В распоряжении гостей собственная ванная комната с ванной',
+  'душевой кабиной и роскошными туалетно-косметическими принадлежностями',
+  'С балкона или террасы всех номеров открывается панорамный вид на Средиземное море.',
+];
+
+
+
+export {TITLE, PRICE, TYPE, TIME, FEATURES, DESCRIPTION};
